Implementing Callbacks in TensorFlow
Write an MNIST classifier that trains to 99% accuracy and stops once this threshold is achieved. In the lecture you saw how this was done for the loss but here you will be using accuracy instead.

Some notes:

Your network should succeed in less than 9 epochs.
When it reaches 99% or greater it should print out the string "Reached 99% accuracy so cancelling training!" and stop training.
If you add any additional variables, make sure you use the same names as the ones used in the class. This is important for the function signatures (the parameters and names) of the callbacks.